fix(auth):调试注册登录页面 feat(channel,topic):增加AI自动编写的页面

+++ internal/utils/jwt.go
@@ -9,12 +9,12 @@ import (
 
 // Claims 自定义声明结构体
 type Claims struct {
-	UserID uint `json:"user_id"`
+	UserID uint64 `json:"user_id"`
 	jwt.StandardClaims
 }
 
 // GenerateToken 生成JWT令牌
-func GenerateToken(userID uint) (string, error) {
+func GenerateToken(userID uint64) (string, error) {
 	cfg := config.GlobalConfig.JWT
 
 	expireTime := time.Now().Add(time.Duration(cfg.ExpireHours) * time.Hour)
